#71 [Refactor] Repository 경로 이동 및 적용
