+++ app/src/main/res/navigation/schedule.xml
@@ -31,7 +31,7 @@
         android:label="ScheduleSelectFragment">
         <argument
             android:name="placeList"
-            app:argType="com.thequietz.travelog.schedule.model.PlaceModel[]" />
+            app:argType="com.thequietz.travelog.schedule.model.SchedulePlaceModel[]" />
         <action
             android:id="@+id/action_scheduleSelectFragment_to_scheduleDetailFragment"
             app:destination="@id/scheduleDetailFragment" />
